+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [Unreleased](https://github.com/warang580/datamix/compare/master...develop)
 
+- Feature: `tap(data, sideEffect)` is part of public API
 - Bugfix: `only()` wasn't working with arrays with dot notation (eg. `['a.b']`)
 - Improved documentation